:bug: Move app to api due to satisfy vercel function pattern match (#2071)
