+++ .github/workflows/backend.yml
@@ -40,18 +40,18 @@ jobs:
       - name: Run ruff format check
         working-directory: backend
         run: |
-          uv run ruff format --check app/
+          uv run ruff format --check api/
 
       - name: Run ruff lint
         working-directory: backend
         run: |
-          uv run ruff check app/
+          uv run ruff check api/
 
       # FIXME: Fix type errors
       # - name: Run type check with mypy
       #   working-directory: backend
       #   run: |
-      #     uv run mypy app/
+      #     uv run mypy api/
 
       - name: Test with pytest
         working-directory: backend
