+++ backend/api/main.py
@@ -2,14 +2,14 @@ import uvicorn
 from fastapi import FastAPI, status
 from fastapi.middleware.cors import CORSMiddleware
 
-from app.constants import API_V1, PRODUCT_NAME, TWO_HOURS_IN_SECONDS
-from app.crud import (
+from api.constants import API_V1, PRODUCT_NAME, TWO_HOURS_IN_SECONDS
+from api.crud import (
     read_accepted_count_by_language_using_user_name,
     read_accepted_count_by_user_name,
     read_longest_streak_by_user_name,
     read_rated_point_sum_by_user_name,
 )
-from app.schemas import (
+from api.schemas import (
     AcceptedCount,
     AcceptedCountByLanguage,
     AtCoderProblemsStatisticsAPI,
@@ -172,4 +172,4 @@ async def read_atcoder_problems_statistics_api(user_name: str):
 
 
 if __name__ == "__main__":
-    uvicorn.run("main:app", host="0.0.0.0", reload=True, port=8000)
+    uvicorn.run("api.main:app", host="0.0.0.0", reload=True, port=8000)
