+++ backend/conftest.py
@@ -1,9 +1,9 @@
+import typing
+
+import pytest
 from fastapi.testclient import TestClient
 
-import typing
-import pytest
-
-from app.main import app
+from api.main import app
 
 
 @pytest.fixture
