+++ backend/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "atcoder-trophies-backend"
-version = "0.3.0"
+version = "0.3.1"
 description = "AtCoder Trophies Backend API server built with FastAPI"
 requires-python = ">=3.12, <3.13"
 
@@ -39,7 +39,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["app"]
+packages = ["api"]
 
 [tool.ruff]
 line-length = 127
